requirements.txt
boto3
dlib
numpy
python-dotenv
pytz
tensorflow==2.3.1
